fixed merge issues
