+++ pkg/onboard/systemd_service_notlinux.go
@@ -15,3 +15,7 @@ func StopSystemdService(serviceName string, skipDeleteDisable, force bool) error
 func GetSystemdServiceStatus(name string) (string, error) {
 	return "", fmt.Errorf("systemd is not available on this platform")
 }
+
+func ResetRestartCounter(service string) error {
+	return fmt.Errorf("systemd is not available on this platform")
+}
